requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
coverage==7.13.5
cryptography==46.0.7
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.135.3
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
motor==3.7.1
multidict==6.7.1
packaging==26.1
pluggy==1.6.0
propcache==0.4.1
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.1
pydantic_core==2.46.1
Pygments==2.20.0
PyJWT==2.12.1
pymongo==4.17.0
pytest==9.0.3
pytest-asyncio==1.3.0
pytest-cov==7.1.0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
requests==2.33.1
rsa==4.9.1
six==1.17.0
starlette==1.0.0
twilio==9.10.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
yarl==1.23.0
